dvmmbr.cpp: Used 32-bit instead of 64-bit type for extended partition type, causing stuff to be rejected.
svn:sync-xref-src-repo-rev: r119387

+++ src/VBox/Runtime/common/dvm/dvmmbr.cpp
@@ -1,4 +1,4 @@
-/* $Id: dvmmbr.cpp 119352 2017-11-30 17:46:06Z knut.osmundsen@oracle.com $ */
+/* $Id: dvmmbr.cpp 119387 2017-12-03 17:06:12Z knut.osmundsen@oracle.com $ */
 /** @file
  * IPRT Disk Volume Management API (DVM) - MBR format backend.
  */
@@ -230,7 +230,7 @@ static void rtDvmFmtMbrDestroy(PRTDVMFMTINTERNAL pThis)
 
 static int rtDvmFmtMbrReadExtended(PRTDVMFMTINTERNAL pThis, PRTDVMMBRENTRY pPrimaryEntry)
 {
-    uint32_t const  cbExt       = pPrimaryEntry->cbPart;
+    uint64_t const  cbExt       = pPrimaryEntry->cbPart;
     uint64_t const  offExtBegin = pPrimaryEntry->offPart;
 
     uint64_t        offCurBegin = offExtBegin;
